requirements.txt
Django==5.0.1
djangorestframework==3.14.0
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
setuptools==69.0.3
django-cors-headers==4.3.1

requests~=2.31.0
